Refactor for containing item and to bring save logic from in memory persistence.

+++ src/main/java/uk/ac/diamond/daq/persistence/logging/ConfigurationLogService.java
@@ -4,11 +4,10 @@ import uk.ac.diamond.daq.persistence.data.LogToken;
 import uk.ac.diamond.daq.persistence.data.PersistableItem;
 import uk.ac.diamond.daq.persistence.service.PersistenceException;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ConfigurationLogService {
-    List<LogToken> listChanges(BigInteger persistenceId) throws PersistenceException;
+    List<LogToken> listChanges(long persistenceId) throws PersistenceException;
 
     LogToken logConfigurationChanges(PersistableItem item, String description) throws PersistenceException;
 }
